backend-chalice/requirements.txt
mypy-boto3-dynamodb==1.34.34
PyJWT==2.7.0
requests==2.31.0
